Remove yarn.lock, add to gitignore
JS land says this file should be checked in, but JS projects have different concerns about fighting with npm to ensure reproducible builds. Perhaps we have that to a small extent as well. I don't think that problem exists to the extent we should allow 4,000 line diffs in our commits though.